+++ client/src/game/shapes/trackers/draw.ts
@@ -2,12 +2,9 @@ import tinycolor from "tinycolor2";
 
 import { TriangulationTarget } from "@/game/visibility/te/pa";
 import { computeVisibility } from "@/game/visibility/te/te";
-import { circleLineIntersection, xyEqual } from "@/game/visibility/te/triag";
 
-import { GlobalPoint, LocalPoint } from "../../geom";
-import { g2l, g2lr, g2lz, getUnitDistance, toRadians } from "../../units";
+import { g2l, g2lr, g2lx, g2ly, toRadians } from "../../units";
 import { Shape } from "../shape";
-import { Circle } from "../variants/circle";
 
 export function drawAuras(shape: Shape, ctx: CanvasRenderingContext2D): void {
     const center = shape.center();
@@ -34,104 +31,44 @@ export function drawAuras(shape: Shape, ctx: CanvasRenderingContext2D): void {
                 g2lr(value),
                 lCenter.x,
                 lCenter.y,
-                g2lr(value + dim),
+                innerRange,
             );
             const tc = tinycolor(aura.colour);
             ctx.fillStyle = gradient;
             gradient.addColorStop(0, aura.colour);
             gradient.addColorStop(1, tc.setAlpha(0).toRgbString());
         }
-        if (!aura.visionSource) {
-            const angleA = toRadians(aura.direction - aura.angle / 2);
-            const angleB = toRadians(aura.direction + aura.angle / 2);
 
-            if (aura.angle < 360) {
-                ctx.moveTo(lCenter.x, lCenter.y);
-                ctx.lineTo(lCenter.x + innerRange * Math.cos(angleA), lCenter.y + innerRange * Math.sin(angleA));
-            }
-            ctx.arc(lCenter.x, lCenter.y, innerRange, angleA, angleB);
-            if (aura.angle < 360) {
-                ctx.lineTo(lCenter.x, lCenter.y);
-            }
-            ctx.fill();
-            ctx.stroke();
-        } else {
+        const angleA = aura.angle === 360 ? 0 : toRadians(aura.direction - aura.angle / 2);
+        const angleB = aura.angle === 360 ? Math.PI * 2 : toRadians(aura.direction + aura.angle / 2);
+
+        // Set visibility polygon as clipping path
+        if (aura.visionSource) {
+            ctx.save();
+
             const polygon = computeVisibility(center, TriangulationTarget.VISION, shape.floor.id);
-            aura.lastPath = updateAuraPath(polygon, center, getUnitDistance(value + dim));
-            try {
-                ctx.fill(aura.lastPath);
-                ctx.stroke(aura.lastPath);
-            } catch (e) {
-                ctx.arc(lCenter.x, lCenter.y, innerRange, 0, 2 * Math.PI);
-                ctx.fill();
-                ctx.stroke();
-                console.warn(e);
-            }
+
+            ctx.beginPath();
+            ctx.moveTo(g2lx(polygon[0][0]), g2ly(polygon[0][1]));
+            for (const point of polygon) ctx.lineTo(g2lx(point[0]), g2ly(point[1]));
+            ctx.clip();
         }
-    }
-}
 
-function updateAuraPath(visibilityPolygon: number[][], center: GlobalPoint, aura: number): Path2D {
-    const auraCircle = new Circle(center, aura);
-    const path = new Path2D();
-    const lCenter = g2l(auraCircle.center());
-    const lRadius = g2lz(auraCircle.r);
-    let firstAngle: number | null = null;
-    let lastAngle: number | null = null;
+        // Draw aura
 
-    const ixs: LocalPoint[][] = [];
+        ctx.beginPath();
 
-    // First find all polygon segments that are actually relevant
-    for (const [i, p] of visibilityPolygon.map((vp) => GlobalPoint.fromArray(vp)).entries()) {
-        const np = GlobalPoint.fromArray(visibilityPolygon[(i + 1) % visibilityPolygon.length]);
-        const pLoc = g2l(p);
-        const npLoc = g2l(np);
-        const ix = circleLineIntersection(auraCircle.center(), auraCircle.r, p, np).map((x) => g2l(x));
-        if (ix.length === 0) {
-            // segment lies completely outside circle
-            if (!auraCircle.contains(p)) continue;
-            // segment lies completely inside circle
-            else ix.push(pLoc, npLoc);
-        } else if (ix.length === 1) {
-            // segment is tangent to circle, segment can be ignored
-            if (xyEqual(ix[0].asArray(), pLoc.asArray()) || xyEqual(ix[0].asArray(), npLoc.asArray())) continue;
-            if (auraCircle.contains(p)) {
-                ix.unshift(pLoc);
-            } else {
-                ix.push(npLoc);
-            }
+        if (aura.angle < 360) {
+            ctx.moveTo(lCenter.x, lCenter.y);
+            ctx.lineTo(lCenter.x + innerRange * Math.cos(angleA), lCenter.y + innerRange * Math.sin(angleA));
         }
-        // Check some bad cases
-        if (ixs.length > 0) {
-            const lastIx = ixs[ixs.length - 1];
-            if (xyEqual(lastIx[0].asArray(), ix[1].asArray()) && xyEqual(lastIx[1].asArray(), ix[0].asArray()))
-                continue;
-            if (xyEqual(lastIx[0].asArray(), ix[0].asArray()) && xyEqual(lastIx[1].asArray(), ix[1].asArray()))
-                continue;
+        ctx.arc(lCenter.x, lCenter.y, innerRange, angleA, angleB);
+        if (aura.angle < 360) {
+            ctx.lineTo(lCenter.x, lCenter.y);
         }
-        ixs.push(ix);
-    }
+        ctx.fill();
+        ctx.stroke();
 
-    if (ixs.length < 1) {
-        path.arc(lCenter.x, lCenter.y, lRadius, 0, 2 * Math.PI);
-        return path;
+        if (aura.visionSource) ctx.restore();
     }
-
-    // If enough interesting edges have been found, cut the circle up.
-    for (const ix of ixs) {
-        const circleHitAngle = Math.atan2(ix[0].y - lCenter.y, ix[0].x - lCenter.x);
-        if (lastAngle === null) {
-            path.moveTo(ix[0].x, ix[0].y);
-            firstAngle = Math.atan2(ix[0].y - lCenter.y, ix[0].x - lCenter.x);
-        } else if (lastAngle !== circleHitAngle) {
-            path.arc(lCenter.x, lCenter.y, lRadius, lastAngle, circleHitAngle);
-        }
-        lastAngle = Math.atan2(ix[1].y - lCenter.y, ix[1].x - lCenter.x);
-        path.lineTo(ix[1].x, ix[1].y);
-    }
-    if (firstAngle && lastAngle) {
-        if (firstAngle !== lastAngle) path.arc(lCenter.x, lCenter.y, lRadius, lastAngle, firstAngle);
-    } else path.arc(lCenter.x, lCenter.y, lRadius, 0, 2 * Math.PI);
-
-    return path;
 }