+++ client/src/game/units.ts
@@ -69,6 +69,7 @@ export function toRadians(degrees: number): number {
     return (degrees * Math.PI) / 180;
 }
 
+// eslint-disable-next-line import/no-unused-modules
 export function toDegrees(radians: number): number {
     return (radians * 180) / Math.PI;
 }